spec/checker: explicit acyclicity for class inheritance
Mirrors how module-level acyclicity is handled. Spec changes:
- §2.2 names the inheritance graph G_{Λ_M} with edge (C, B) iff
  Λ_M(C) = (\vec{x}, B).
- Module rule (Fig 6) gains the premise 'G_{Λ_M} acyclic' alongside
  the existing imports-resolved and class-name-uniqueness premises.
- fields_M is now a well-defined least solution rather than relying
  on non-termination to implicitly reject cyclic inheritance.

Checker changes:
- build_class_context performs a DFS cycle check on the inheritance
  graph; raises IllFormedModule with a new CyclicInheritance reason
  on detection. Previously the checker would RecursionError.
- has_cycle moves to check_module.py so it can be shared.

Plus a new ill-formed/semantic test for cyclic inheritance.
