+++ PurePy-spec.tex
@@ -262,6 +262,7 @@ which follow Python.
 Class contexts $\Lambda$ are finite maps from class names to pairs $(\vec{x}, C)$, where $\vec{x}$ is a sequence of distinct
 field names (in declaration order) and $C$ names the base class (or $\bot$ for no base). Metavariables $C$ and $B$ range over class names.
 Union of class contexts is undefined when the operands disagree on a shared key.
+The \emph{inheritance graph} $G_{\Lambda_M}$ of a class context has an edge $(C, B)$ iff $\Lambda_M(C) = (\vec{x}, B)$ for some $\vec{x}$.
 
 \subsection{Modules and entry point}
