+++ test/module-level/ill-formed/semantic/cyclic_inheritance.error.expected
@@ -1,0 +1,1 @@
+cyclic inheritance
