+++ test/module-level/ill-formed/semantic/cyclic_inheritance.exception.expected
@@ -1,0 +1,1 @@
+NameError
