+++ test/module-level/ill-formed/semantic/cyclic_inheritance.py
@@ -1,0 +1,11 @@
+# rule: class-extend
+from dataclasses import dataclass
+from typing import Any
+
+@dataclass
+class A(B):  # PurePy: error (cyclic inheritance); Python: NameError
+    x: Any
+
+@dataclass
+class B(A):
+    y: Any
